wiring up iphone landing page, just for dev for now
