+++ kifi-backend/modules/shoebox/app/views/marketing/iPhoneLanding.scala.html
@@ -1,19 +1,19 @@
 <!doctype html>
 <meta charset="utf-8">
-<meta name="viewport" content="width=device-width,user-scalable=no">
-<link rel="stylesheet" href="landing.css">
-<title>Get Kifi</title>
-<body>
-<div id="k-iscroll">
+<meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=0">
+<link rel="shortcut icon" type="image/png" href="/assets/images/kifi-favi-50.png">
+<link rel="stylesheet" href="/assets/m/s/iphone-landing.css">
+<title>Kifi • Search like normal, find like never before</title>
+<main>
   <section class="k-top">
-    <img class="k-logo" src="kifi-logo-white.png" alt="kifi">
+    <img class="k-logo" src="/assets/m/i/kifi-logo-white.png" alt="kifi">
     <h1 class="k-h1">Upgrade your mobile browsing with the social library</h1>
     <ul class="k-why">
-      <li>Easily keep pages you <img class="k-heart" src="heart.png" alt="love"></li>
+      <li>Easily keep pages you <img class="k-heart" src="/assets/m/i/heart.png" alt="love"></li>
       <li>Discover what your friends kept too</li>
       <li>Share &amp; Chat right from your browser</li>
     </ul>
-    <img class="k-iphone k-out" src="held-iphone.png" alt="iPhone running Kifi">
+    <img class="k-iphone k-out" src="/assets/m/i/held-iphone.png" alt="iPhone running Kifi">
     <a class="k-download" href="javascript:">Download Kifi</a>
   </section>
   <section class="k-card k-card-1">
@@ -37,7 +37,5 @@
     running Chrome or<br>Firefox. Android and iPad<br>
     coming soon
   </section>
-</div>
-</body>
-<script src="iscroll-probe.js"></script>
-<script src="landing.js"></script>
+</main>
+<script src="/assets/m/j/iphone-landing.js"></script>
