+++ kifi-backend/modules/shoebox/app/views/marketing/mobileLanding.scala.html
@@ -27,8 +27,6 @@
 
     <link rel="shortcut icon" type="image/png" href="/assets/images/kifi-favi-50.png">
     <link href="//fonts.googleapis.com/css?family=Raleway:100,200,300,400|Open+Sans:100,200,300,400" rel="stylesheet" type="text/css">
-    <link rel="stylesheet" href="/assets/css/normalize.css">
-
   <link href="//fonts.googleapis.com/css?family=Open+Sans:100,200,300,400" rel="stylesheet" type="text/css">
   <link rel="stylesheet" href="/assets/css/normalize.css">
   <link rel="stylesheet" href="/assets/css/landing-mobile.css">
