+++ kifi-backend/modules/shoebox/public/m/j/iphone-landing.js
@@ -1,3 +1,4 @@
+/*! iscroll-probe.js v5.1.1 ~ (c) Matteo Spinelli ~ cubiq.org/license */
 (function(g,n,f){function p(a,b){this.wrapper="string"==typeof a?n.querySelector(a):a;this.scroller=this.wrapper.children[0];this.scrollerStyle=this.scroller.style;this.options={resizeScrollbars:!0,mouseWheelSpeed:20,snapThreshold:0.334,startX:0,startY:0,scrollY:!0,directionLockThreshold:5,momentum:!0,bounce:!0,bounceTime:600,bounceEasing:"",preventDefault:!0,preventDefaultException:{tagName:/^(INPUT|TEXTAREA|BUTTON|SELECT)$/},HWCompositing:!0,useTransition:!0,useTransform:!0};for(var c in b)this.options[c]=
 b[c];this.translateZ=this.options.HWCompositing&&d.hasPerspective?" translateZ(0)":"";this.options.useTransition=d.hasTransition&&this.options.useTransition;this.options.useTransform=d.hasTransform&&this.options.useTransform;this.options.eventPassthrough=!0===this.options.eventPassthrough?"vertical":this.options.eventPassthrough;this.options.preventDefault=!this.options.eventPassthrough&&this.options.preventDefault;this.options.scrollY="vertical"==this.options.eventPassthrough?!1:this.options.scrollY;
 this.options.scrollX="horizontal"==this.options.eventPassthrough?!1:this.options.scrollX;this.options.freeScroll=this.options.freeScroll&&!this.options.eventPassthrough;this.options.directionLockThreshold=this.options.eventPassthrough?0:this.options.directionLockThreshold;this.options.bounceEasing="string"==typeof this.options.bounceEasing?d.ease[this.options.bounceEasing]||d.ease.circular:this.options.bounceEasing;this.options.resizePolling=void 0===this.options.resizePolling?60:this.options.resizePolling;
@@ -60,3 +61,27 @@ this.maxPosY/this.scroller.maxScrollY);this.updatePosition()},updatePosition:fun
 3*(b-this.maxPosY),8),this.indicatorStyle.height=this.height+"px",b=this.maxPosY+this.indicatorHeight-this.height):b=this.maxBoundaryY:"scale"==this.options.shrink&&this.height!=this.indicatorHeight&&(this.height=this.indicatorHeight,this.indicatorStyle.height=this.height+"px"));this.x=a;this.y=b;this.scroller.options.useTransform?this.indicatorStyle[d.style.transform]="translate("+a+"px,"+b+"px)"+this.scroller.translateZ:(this.indicatorStyle.left=a+"px",this.indicatorStyle.top=b+"px")},_pos:function(a,
 b){0>a?a=0:a>this.maxPosX&&(a=this.maxPosX);0>b?b=0:b>this.maxPosY&&(b=this.maxPosY);a=this.options.listenX?f.round(a/this.sizeRatioX):this.scroller.x;b=this.options.listenY?f.round(b/this.sizeRatioY):this.scroller.y;this.scroller.scrollTo(a,b)},fade:function(a,b){if(!b||this.visible){clearTimeout(this.fadeTimeout);this.fadeTimeout=null;var c=a?250:500,e=a?0:300;this.wrapperStyle[d.style.transitionDuration]=c+"ms";this.fadeTimeout=setTimeout(function(a){this.wrapperStyle.opacity=a;this.visible=+a}.bind(this,
 a?"1":"0"),e)}}};p.utils=d;"undefined"!=typeof module&&module.exports?module.exports=p:g.IScroll=p})(window,document,Math);
+
+setTimeout(function initScrolling() {
+  var $top = document.querySelector('.k-top');
+  var $topPar = $top.parentNode;
+  var fixed = false;
+  var iScr = new IScroll(document.body, {probeType: 3, disableMouse: true, disablePointer: true});
+  iScr.on('scroll', fixTop);
+  iScr.on('scrollEnd', fixTop);
+  function fixTop(e) {
+    var y = this.y << 0;
+    if ((y < -367) !== fixed) {
+      fixed = !fixed;
+      $top.parentNode.removeChild($top);
+      $top.style.position = fixed ? 'fixed' : '';
+      $top.style.top = fixed ? '-367px' : '';
+      var p = fixed ? $topPar.parentNode : $topPar;
+      p.insertBefore($top, p.firstChild);
+      setTimeout(iScr.refresh.bind(iScr), 0);
+    }
+  }
+  document.addEventListener('touchmove', function (e) { e.preventDefault() }, false);
+}, 100);
+
+document.querySelector('.k-iphone').classList.remove('k-out');
